feat: academicRecord 호출 훅 이름 변경 및 createFormData 이용한 호출 로직 단순화 (#146)
